Bubble up the invalid field in the EncodeError
We often get invalid messages and the error does not say anything about
which field was invalid. This patch adds a validate! method to Message
and Message's instance variables that throws an exception rather than
returning false like Message#valid? . It also adds the invalid field to
the EncodeError so that it can be inspected.